[default] feat: TC 추가, 리팩토링, 등
