+++ src/main/java/com/hhplus/course/lecture/domain/LectureId.java
