+++ src/main/java/com/hhplus/course/lecture/infra/repository/JpaLectureRepository.java
