+++ src/main/java/com/hhplus/course/lecture/presentation/controller/dto/LectureResponse.java
@@ -1,0 +1,17 @@
+package com.hhplus.course.lecture.presentation.controller.dto;
+
+import com.hhplus.course.lecture.domain.Lecture;
+
+public record LectureResponse(
+        String lectureId,
+        String title,
+        String lecturer
+) {
+    public static LectureResponse of(Lecture lecture) {
+        return new LectureResponse(
+                lecture.getId().getId(),
+                lecture.getTitle(),
+                lecture.getLecturer().getName()
+        );
+    }
+}
