+++ src/main/java/com/hhplus/course/register/presentation/dto/ApplyRequest.java
@@ -1,0 +1,11 @@
+package com.hhplus.course.register.presentation.dto;
+
+import java.time.LocalDate;
+
+public record ApplyRequest(
+        String lectureId,
+        String userId,
+        LocalDate date
+) {
+}
+
