+++ src/main/resources/application.properties
@@ -1,1 +1,10 @@
 spring.application.name=course
+
+spring.jpa.database=postgresql
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.datasource.url=jdbc:postgresql://localhost:5432/hhplus
+spring.datasource.username=postgres
+spring.datasource.password=
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.jpa.properties.hibernate.default_schema=course
