+++ src/test/java/com/hhplus/course/lecture/LectureTest.java
@@ -1,0 +1,7 @@
+package com.hhplus.course.lecture;
+
+public class LectureTest {
+
+
+
+}
