--- src/test/java/com/hhplus/course/lecture/domain/LectureTest.java
@@ -1,5 +1,0 @@
-package com.hhplus.course.lecture.domain;
-
-public class LectureTest {
-
-}
