--- src/test/java/com/hhplus/course/user/domain/UserId.java
@@ -1,26 +1,0 @@
-package com.hhplus.course.user.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-import static lombok.AccessLevel.PROTECTED;
-
-@Embeddable
-@Getter
-@EqualsAndHashCode
-@NoArgsConstructor(access = PROTECTED)
-@AllArgsConstructor
-public class UserId implements Serializable {
-    @Column(name = "users_id")
-    String id;
-
-    public static UserId of(String id) {
-        return new UserId(id);}
-
-}
