+++ src/test/resources/application-test.yaml
@@ -1,0 +1,19 @@
+spring:
+  datasource :
+    url: jdbc:postgresql://localhost:5432/hhplus
+#    driverClassName: org.h2.Driver
+    username: postgres
+    password:
+
+  jpa:
+    hibernate:
+      ddl-auto: update
+
+    properties:
+      hibernate:
+        default_schema: course-test
+        #    database-platform: org.hibernate.dialect.H2Dialect
+
+logging:
+  level:
+    org.hibernate.orm.jdbc.bind: trace
